Added basic multi-project app implementation.
This adds a basic but functional app, which makes use of a multi-project structure. This is just a replication of a production app that I maintain, so might seem slightly strange given how empty it is.

I've tried to include some basic example code where needed (not only to clarify the use of each project, but also to make sure each has something to compile), but have tried to keep it quite minimal.